requirements.txt
streamlit>=1.28.0
transformers>=4.35.0
torch>=2.1.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
python-docx>=0.8.11
tqdm>=4.66.0  # For progress bars